// types.ts
export enum TaskType {
  HOMEWORK = 'Tema de Casa',
  PROJECT = 'Trabalho',
  EXAM = 'Prova'
}

export enum Subject {
  MATH = 'Matemática',
  PORTUGUESE = 'Português',
  SCIENCE = 'Ciências',
  HISTORY = 'História',
  GEOGRAPHY = 'Geografia',
  ENGLISH = 'Inglês',
  ARTS = 'Artes',
  PE = 'Educação Física'
}

export interface Task {
  id: string;
  title: string;
  type: TaskType;
  subject: Subject;
  date: string; // ISO string YYYY-MM-DD
  description?: string;
  examContent?: string; // Specific for exams
  isCompleted: boolean;
  aiStudyPlan?: string; // Generated by Gemini
  attachment?: string; // Firebase Storage URL
}

export const SUBJECT_COLORS: Record<Subject, string> = {
  [Subject.MATH]: 'from-blue-400 to-cyan-400',
  [Subject.PORTUGUESE]: 'from-pink-500 to-rose-400',
  [Subject.SCIENCE]: 'from-emerald-400 to-teal-400',
  [Subject.HISTORY]: 'from-amber-400 to-orange-400',
  [Subject.GEOGRAPHY]: 'from-violet-500 to-purple-400',
  [Subject.ENGLISH]: 'from-indigo-400 to-blue-500',
  [Subject.ARTS]: 'from-fuchsia-500 to-pink-400',
  [Subject.PE]: 'from-orange-500 to-red-400'
};

export const SUBJECT_ABBREVIATIONS: Record<Subject, string> = {
  [Subject.MATH]: 'MAT',
  [Subject.PORTUGUESE]: 'PORT',
  [Subject.SCIENCE]: 'CIÊN',
  [Subject.HISTORY]: 'HIST',
  [Subject.GEOGRAPHY]: 'GEO',
  [Subject.ENGLISH]: 'ING',
  [Subject.ARTS]: 'ART',
  [Subject.PE]: 'ED.F'
};

export const SUBJECT_BORDER_COLORS: Record<Subject, string> = {
  [Subject.MATH]: 'border-blue-400',
  [Subject.PORTUGUESE]: 'border-pink-500',
  [Subject.SCIENCE]: 'border-emerald-400',
  [Subject.HISTORY]: 'border-amber-400',
  [Subject.GEOGRAPHY]: 'border-violet-500',
  [Subject.ENGLISH]: 'border-indigo-400',
  [Subject.ARTS]: 'border-fuchsia-500',
  [Subject.PE]: 'border-orange-500'
};

export const SUBJECT_TEXT_COLORS: Record<Subject, string> = {
    [Subject.MATH]: 'text-blue-400',
    [Subject.PORTUGUESE]: 'text-pink-500',
    [Subject.SCIENCE]: 'text-emerald-400',
    [Subject.HISTORY]: 'text-amber-400',
    [Subject.GEOGRAPHY]: 'text-violet-500',
    [Subject.ENGLISH]: 'text-indigo-400',
    [Subject.ARTS]: 'text-fuchsia-500',
    [Subject.PE]: 'text-orange-500'
};


export const TYPE_ICONS: Record<TaskType, string> = {
  [TaskType.HOMEWORK]: '📝',
  // FIX: Used TaskType enum instead of Subject for PROJECT and EXAM keys.
  [TaskType.PROJECT]: '🚀',
  [TaskType.EXAM]: '🏆'
};